requirements.txt
aiofiles==23.2.1
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.12.1
attrs==25.4.0
certifi==2026.2.25
charset-normalizer==3.4.4
click==8.3.1
dataclasses-json==0.6.7
distro==1.9.0
fastapi==0.110.0
frozenlist==1.8.0
greenlet==3.3.2
groq==0.37.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
Jinja2==3.1.3
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.1.20
langchain-community==0.0.38
langchain-core==0.1.53
langchain-groq==0.1.3
langchain-text-splitters==0.0.2
langsmith==0.1.147
MarkupSafe==3.0.3
marshmallow==3.26.2
multidict==6.7.1
mypy_extensions==1.1.0
neo4j==5.17.0
newsapi-python==0.2.7
numpy==1.26.4
orjson==3.11.7
packaging==23.2
propcache==0.4.1
pydantic==2.7.0
pydantic_core==2.18.1
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2025.2
PyYAML==6.0.3
regex==2026.2.28
requests==2.31.0
requests-toolbelt==1.0.0
sniffio==1.3.1
SQLAlchemy==2.0.47
starlette==0.36.3
tenacity==8.5.0
tiktoken==0.7.0
typing-inspect==0.9.0
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.29.0
Wikipedia-API==0.6.0
yarl==1.23.0
